--- templates/register.html
@@ -1,8 +1,0 @@
-{% extends 'base.html' %}
-{% block title %}
-    GUATETRIVIA - SIGNUP
-{% endblock %}
-
-{% block content %}
-
-{% endblock %}
