--- trivia.py
@@ -1,30 +1,0 @@
-from flask import Flask, render_template
-from flask_sqlalchemy import SQLAlchemy
-
-app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///guatetrivia.db'
-db = SQLAlchemy(app)
-
-
-class User(db.Model):
-    id = db.Column(db.Integer(), primary_key=True)
-    first_name = db.Column(db.String(length=50), nullable=False)
-    last_name = db.Column(db.String(length=50), nullable=False)
-    username = db.Column(db.String(length=20), nullable=False, unique=True)
-    email = db.Column(db.String(length=50), nullable=False, unique=True)
-    password = db.Column(db.String(length=50), nullable=False)
-
-
-@app.route('/')
-def home_page():
-    return render_template('home.html')
-
-
-@app.route('/login')
-def login():
-    return render_template('login.html')
-
-
-@app.route('/register')
-def register():
-    return render_template('register.html')
