+++ trivia/forms.py
@@ -1,0 +1,11 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, SubmitField
+from wtforms.validators import DataRequired, Length, Email
+
+
+class RegisterForm(FlaskForm):
+    username = StringField(label='username')
+    email_address = StringField(label='email')
+    password1 = PasswordField(label='password')
+    password2 = PasswordField(label='repeat password')
+    submit = SubmitField(label='submit')
