+++ trivia/templates/base.html
@@ -36,6 +36,8 @@
     </nav>
 
     <h1>Bienvenido a GuateTrivia</h1>
+    {% block content %}
+    {% endblock %}
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
   </body>
 </html>
